Restructure lib/events
