+++ lib/events/src/index.ts
@@ -1,1 +1,1 @@
-export * from './';
+export * from './events';
